move to constant
